requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyPDF2==3.0.1
typing_extensions==4.12.2
waitress==3.0.2
Werkzeug==3.1.3
zipp==3.21.0
python-docx==1.1.2
tqdm==4.67.1
Pillow==10.2.0
